refactor(Name, Names) : 도메인 패키지 변경
- 기존 : common 패키지
- 변경 : player 패키지
