+++ src/main/java/BlackjackController.java
@@ -1,6 +1,6 @@
 import blackjack.domain.Blackjack;
 import blackjack.domain.card.Card;
-import blackjack.domain.common.Names;
+import blackjack.domain.player.Names;
 import blackjack.domain.player.Dealer;
 import blackjack.domain.player.GamePlayer;
 import blackjack.domain.player.Players;
