+++ src/main/java/blackjack/domain/player/GamePlayer.java
@@ -2,7 +2,6 @@ package blackjack.domain.player;
 
 
 import blackjack.domain.card.Card;
-import blackjack.domain.common.Name;
 import blackjack.domain.card.Cards;
 import blackjack.domain.result.ResultStatus;
 import java.util.List;
