+++ src/test/java/blackjack/domain/common/NamesTest.java
@@ -2,6 +2,7 @@ package blackjack.domain.common;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import blackjack.domain.player.Names;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
